# backend-flask/api/open_api/openapi.yaml
openapi: 3.0.3
info:
  title: "SC API pos"
  description: "Модуль для интеграции с кассами для лояльности на смарт контрактах."
  version: 1.0.0
servers:
  - url: https://dm.api/20130701/
    description: "Production сервер"
  - url: http://staging-api.example.com
    description: "Staging сервер"
...
paths:
...
  /loyalties/{merchant_id}:
    get:
      summary: "Поиск лояльности"
      operationId: src.loyalties.views.get_loyalty
      tags:
      - Loyalties
      description: "Получение полной информации о ПЛ (Программы Лояльности) участника Системы Rusbonus. Программа лояльности может отсутствовать, а также содержать нулевые значения порогов и скидок\n\n."
      parameters:
      - $ref: '#/components/parameters/DM-Authorization'
      - $ref: '#/components/parameters/_dmapptoken'
      - $ref: '#/components/parameters/merchant_id'
      responses:
        '200':
          description: ""
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Loyalty'
  /loyalties/{merchant_id}/dmcard/{card}:
    get:
      summary: "Поиск лояльности"
      operationId: src.loyalties.views.get_loyalty_by_card
      tags:
      - Loyalties
      description: "Получение полной информации о ПЛ (Программы Лояльности) участника Системы Rusbonus. Программа лояльности может отсутствовать, а также содержать нулевые значения порогов и скидок\n\n."
      security:
      - bearerAuthPos: []
      parameters:
      - $ref: '#/components/parameters/DM-Authorization'
      - $ref: '#/components/parameters/_dmapptoken'
      - $ref: '#/components/parameters/merchant_id'
      - $ref: '#/components/parameters/card'
      responses:
        '200':
          description: ""
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Loyalty'
  /evotor/user:
    post:
      summary: "Регистрация учётной записи"
      operationId: src.evotor.views.callback_create_user
      tags:
      - Evotor
      description: "Callback о регистрация учётной записи в системе эвотор"
      security:
        - bearerAuthEvotor: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                  maxLength: 150
                  description: "Идентификатор пользователя Эвотор."
                  example: "01-000000000000001"
                username:
                  $ref: "#/components/schemas/email"
                password:
                  type: string
                  maxLength: 50
                  description: "Пароль пользователя"
                  example: "12345678"
                password_confirm:
                  type: string
                  maxLength: 50
                  description: "Пароль пользователя ещё раз"
                  example: "12345678"
                org_name:
                  type: string
                  maxLength: 50
                  description: "Название организации"
                  example: "ИП Иванов"

      responses:
        '200':
          description: ""
          content:
            application/json:
              schema:
                properties:
                  userId:
                    type: string
                    maxLength: 150
                    description: "Идентификатор пользователя Эвотор."
                    example: "01-000000000000001"
                  token:
                    type: string
                    description: "Bearer-токен, с помощью которого Облако Эвотор авторизует свои запросы к стороннему сервису."
                    example: "toaWaep4chou7ahkoogiu9Iusaht9ima"
  /evotor/user/auth:
    post:
      summary: "Авторизация пользователя"
      operationId: src.evotor.views.callback_verify_user
      tags:
      - Evotor
      description: "Callback о авторизации существующего пользователя в нашем сервисе с помощью данных учётной записи пользователя Эвотор"
      security:
        - bearerAuthEvotor: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                  maxLength: 150
                  description: "Идентификатор пользователя Эвотор."
                  example: "01-000000000000001"
                username:
                  $ref: "#/components/schemas/email"
                password:
                  type: string
                  maxLength: 50
                  description: "Пароль пользователя в стороннем сервисе"
                  example: "12345678"

      responses:
        '200':
          description: ""
          content:
            application/json:
              schema:
                properties:
                  userId:
                    type: string
                    maxLength: 150
                    description: "Идентификатор пользователя Эвотор."
                    example: "01-000000000000001"
                  token:
                    type: string
                    description: "Bearer-токен, с помощью которого Облако Эвотор авторизует свои запросы к стороннему сервису."
                    example: "toaWaep4chou7ahkoogiu9Iusaht9ima"
  /evotor/user/token:
    post:
      summary: "Передача токена Облака Эвотор"
      operationId: src.evotor.views.callback_token
      tags:
      - Evotor
      description: "Передаёт токен Облака Эвотор, после того, как пользователь установил приложение в Личном кабинете. Используйте токен для авторизации запросов своего приложения к Облаку Эвотор."
      security:
        - bearerAuthEvotor: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                  maxLength: 150
                  description: "Идентификатор пользователя Эвотор."
                token:
                  type: string
                  description: "Токен, с которым приложение будет обращаться в Облако Эвотор для запроса данных по пользователю с идентификатором."
                  example: "f46b89a5-8e80-4591-b0aa-94551790444b"

      responses:
        '200':
          description: ""
          content:
            {}
  /evotor/subscription/billing:
    post:
      summary: "События подписки"
      operationId: src.evotor.views.callback_subscription
      tags:
      - Evotor
      description: "Связанные с биллингом события, которые Облако Эвотор передаёт в сторонний сервис."
      security:
        - bearerAuthEvotor: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                  maxLength: 150
                  description: "Идентификатор пользователя Эвотор."
                subscriptionId:
                  type: string
                  description: "Идентификатор подписки. Формат – uuid4."
                  example: "a99fbf70-6307-4acc-b61c-741ee9eef6c0"
                productId:
                  type: string
                  description: "Идентификатор приложения. Формат – uuid4."
                  example: "c0d01x35-5193-4cc2-9bfb-be20e0679498"
                timestamp:
                  type: string
                  format: date-time
                  description: "Дата и время отправки события. В соответствовии с ISO 8601."
                  example: "2017-04-20T18:26:37.753+0000"
                sequenceNumber:
                  type: integer
                  description: "Номер события в последовательности. Номер непрерывно возрастает начиная с единицы. Необходим для соблюдения порядка обработки событий. Номер события уникален в рамках подписки (subscriptionId), таким образом, при переустановке приложения номерация событий начнётся сначала."
                  example: 4
                type:
                  type: string
                  description: "Типы событий:\n\nSubscriptionCreated – новая подписка. Сообщает о том, что пользователь установил приложение в Личном кабинете. Приходит в начале пробного периода или перед сообщением об успешной оплате, если пробного периода нет.\n\nAddonsUpdated – список платных опций, выбранных пользователем.\n\nSubscriptionActivated – подписка активирована. Сообщает об успешной оплате.\n\nSubscriptionRenewed – подписка продлена на следующий период. Сообщает об успешной оплате очередного периода.\n\nSubscriptionTermsChanged – изменились условия подписки, например, тарифный план или количество устройств.\n\nSubscriptionTerminationRequested – Пользователь отправил запрос на завершение подписки (удалил приложение из Личного кабинета). Пользователь может возобновить подписку до окончания оплаченного периода.\n\nSubscriptionTerminated – Подписка завершена. Приходит если не прошла регулярная оплата, независимо от того запросил пользователь завершение подписки или нет."
                  example: "a99fbf70-6307-4acc-b61c-741ee9eef6c0"

      responses:
        '200':
          description: ""
          content:
            {}
  /evotor/inventories/stores/{storeUuid}/documents:
    put:
      summary: "События магазина"
      operationId: src.evotor.views.callback_shops
      tags:
      - Evotor
      description: "Передать документы в сторонний сервис."
      parameters:
        - in: path
          name: storeUuid
          schema:
            type: string
          required: true

      security:
        - bearerAuthEvotor: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      uuid:
                        type: string
                        description: "Идентификатор приложения. Формат – uuid4."
                        example: "c0d01x35-5193-4cc2-9bfb-be20e0679498"
                      address:
                        type: string
                        description: "Адрес магазина."
                        example: "Адрес"
                      name:
                        type: string
                        description: "Название магазина"
                        example: "Сокольники"

      responses:
        '200':
          description: ""
          content:
            {}

  # Авторизация. Передать: почта и пароль. Вернуть: да, нет.
  # Получить информацию о лояльности. Передать: токен. Вернуть: инфу о лояльности.
  # Измение лояльности. Передать: параметры которые изменяются. Вернуть: да, либо ошибку подробную.
...
  /adminsc/loyalty:
    get:
      summary: "Информация о лояльности"
      operationId: src.adminsc.views.get_loyalty
      tags:
      - AdminSC
      description: "Информация о лояльности"
      security:
        - bearerAuthAdmin: []
      responses:
        '200':
          description: ""
          content:
            application/json:
              schema:
                properties:
                  accrualPercent:
                    type: string
                    description: "курс начисления бонусных баллов"
                    example: "1"
                  minAccrualThreshold:
                    type: string
                    description: "минимальная сумма покупки, при которой производится начисление бонусных баллов"
                    example: "1"
                  accrualRate:
                    type: string
                    description: "курс списания бонусного балла"
                    example: "1"
                  maxWriteoffPercent:
                    type: string
                    description: "процент покупки, который может быть оплачен за счет бонусов"
                    example: "100"
                  bonusActivationDelay:
                    type: string
                    description: "срок активации бонусных баллов, сек."
                    example: "400"
                  bonusLifetime:
                    type: string
                    description: "срок жизни бонусных баллов, сек."
                    example: "500"
                  name:
                    type: string
                    description: "Тип контракта."
                    example: "Bonus"
                  generation:
                    type: string
                    description: "Версия контракта"
                    example: "3.0"
                  isSimultaneous:
                    type: boolean
                    description: "одновременное списание и начисление баллов."
                    example: true
              examples:
                v3:
                  value:
                    accrualPercent: "2"
                    minAccrualThreshold: "1"
                    accrualRate: "1"
                    maxWriteoffPercent: "100"
                    isSimultaneous: true
                    name: "Bonus"
                    generation: "3.0"
                v2:
                  value:
                    accrualPercent: "2"
                    minAccrualThreshold: "1"
                    accrualRate: "1"
                    maxWriteoffPercent: "100"
                    bonusActivationDelay: "400"
                    bonusLifetime: "500"
                    isSimultaneous: true
                    name: "Bonus"
                    generation: "2.0"

    put:
      summary: "Редактировать лояльность"
      operationId: src.adminsc.views.edit_loyalty
      tags:
      - AdminSC
      description: "Редактировать лояльность"
      security:
        - bearerAuthAdmin: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                accrualPercent:
                    type: string
                    description: "курс начисления бонусных баллов"
                    example: "100"
                minAccrualThreshold:
                  type: string
                  description: "минимальная сумма покупки, при которой производится начисление бонусных баллов"
                  example: "1"
                accrualRate:
                  type: string
                  description: "курс списания бонусного балла"
                  example: "1"
                maxWriteoffPercent:
                  type: string
                  description: "процент покупки, который может быть оплачен за счет бонусов"
                  example: "100"
                bonusActivationDelay:
                  type: string
                  description: "срок активации бонусных баллов, сек."
                  example: "1"
                bonusLifetime:
                  type: string
                  description: "срок жизни бонусных баллов, сек."
                  example: "999999"
                isSimultaneous:
                  type: boolean
                  description: "одновременное списание и начисление баллов."
                  example: true

      responses:
        '202':
          description: ""
          content:
            {}
        '402':
          description: "Требуется оплата. Подписка не актива"
          content:
            application/json:
              schema:
                example: {}
components:
  securitySchemes:
    bearerAuthEvotor:
      type: http
      scheme: bearer
      bearerFormat: JWT
      x-bearerInfoFunc: src.middleware.bearer_auth_evotor

    bearerAuthAdmin:
      type: http
      scheme: bearer
      bearerFormat: JWT
      x-bearerInfoFunc: src.middleware.bearer_auth_adminsc

    bearerAuthPos:
      type: apiKey
      in: header
      name: Dmpos-Authorization
      x-apikeyInfoFunc: src.middleware.bearer_auth_pos
...
  parameters:
    DM-Authorization:
      name: "DM-Authorization"
      in: header
      schema:
        type: string
        example: "dmapptoken 550e8400-e29b-41d4-a716-446655440000"
        description: "Этот ключ создаётся и выдаётся разработчикам приложений, использующим API, сотрудниками компании Rusbonus."
    _dmapptoken:
      name: "_dmapptoken"
      in: query
      schema:
        type: string
        example: "dmapptoken 550e8400-e29b-41d4-a716-446655440000"
        description: "Этот ключ создаётся и выдаётся разработчикам приложений, использующим API, сотрудниками компании Rusbonus."
    merchant_id:
      name: "merchant_id"
      in: path
      required: true
      schema:
        type: string
        maxLength: 100
...
    card:
      name: "card"
      in: path
      required: true
      schema:
        type: string
        maxLength: 100
...
  schemas:
...
    email:
      type: string
      maxLength: 100
      description: "Строка цифр длиной до 100.\n\nАдрес электронной почты нового пользователя."
      example: "test@example.com"
...
    Loyalty:
      type: object
      properties:
        type:
          type: string
          example: "bonus"
        currency_code:
          type: integer
          format: int64
          example: 643
          description: "Целое положительное число\n\nЦифровой код валюты в стандарте ISO 4217 [http://ru.wikipedia.org/wiki/ISO_4217]."
        currency_name:
          type: string
          example: "RUB"
          description: "Строка символов длиной 3.\n\nАлфавитный код валюты в стандарте ISO 4217 [http://ru.wikipedia.org/wiki/ISO_4217]."
        min_purchase_amount:
          type: integer
          format: int64
          example: 1
          description: "Положительное целое число.\n\nМинимальная сумма покупки для начисления бонусных баллов."
        max_purchase_percentage:
          type: integer
          format: int64
          example: 50
          description: "Положительное целое число в диапазоне от 0 до 100.\n\nМаксимальная часть стоимости покупки в процентах, которая может быть оплачена бонусными баллами."
        expiration:
          type: integer
          format: int64
          example: 30
          description: "Положительное целое число.\n\nКоличество дней, спустя которые начисленные бонусные баллы сгорят."
        amount_to_bonus:
          type: array
          description: "Список алфавитно-цифровых строк без знаков препинания, разделённых запятыми.\n\nСписок идентификаторов купонов (сертификатов) пользователя, которые следует погасить в этой покупке."
          items:
            type: string
          example: [100, "10.0000"]
        bonus_to_amount:
          type: array
          description: "Список алфавитно-цифровых строк без знаков препинания, разделённых запятыми.\n\nСписок идентификаторов купонов (сертификатов) пользователя, которые следует погасить в этой покупке."
          items:
            type: string
          example: [1, "0.1000"]
...